+++ src/main/java/br/com/zup/ecommerce/controllers/ProdutoAtualizacaoController.java
@@ -6,7 +6,6 @@ import br.com.zup.ecommerce.entities.produto.opiniao.OpiniaoProduto;
 import br.com.zup.ecommerce.entities.produto.opiniao.OpiniaoProdutoNovoRequest;
 import br.com.zup.ecommerce.entities.produto.pergunta.PerguntaProduto;
 import br.com.zup.ecommerce.entities.produto.pergunta.PerguntaProdutoNovoRequest;
-import br.com.zup.ecommerce.service.email.EnviarEmail;
 import br.com.zup.ecommerce.service.produto.AtualizacaoProduto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
@@ -15,7 +14,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -24,7 +22,7 @@ import javax.validation.Valid;
 import java.util.Set;
 
 /**
- * Contagem de carga intrínseca da classe: 8
+ * Contagem de carga intrínseca da classe: 7
  */
 
 @RestController
@@ -38,10 +36,6 @@ public class ProdutoAtualizacaoController {
     //1
     private AtualizacaoProduto atualizacaoProduto;
 
-    @Autowired
-    //1
-    private EnviarEmail enviarEmail;
-
     @PostMapping("/imagens")
     @Transactional
     //1
@@ -50,10 +44,10 @@ public class ProdutoAtualizacaoController {
         //1
         Produto produto = atualizacaoProduto.getProduto(id,manager);
 
-        //Lança uma ResponseStatusException se usuário inválido
+        //Lança uma ResponseStatusException se usuário for inválido
         atualizacaoProduto.validaDonoProduto(manager, produto, user);
 
-        Set<String> links = atualizacaoProduto.enviaImagem(novasImagens);
+        Set<String> links = atualizacaoProduto.getAtualizacaoImagem().enviaImagem(novasImagens);
         produto.incluirImagens(links);
 
         manager.merge(produto);
@@ -89,9 +83,8 @@ public class ProdutoAtualizacaoController {
 
         manager.merge(produto);
 
-        String assunto = montarAssuntoEmailPergunta(produto);
-        String mensagem = montarEmailPergunta(produto, pergunta);
-        enviarEmail.enviarEmail(produto.getDono().getLogin(), assunto, mensagem);
+        //Envia e-mail da pergunta para o dono do produto
+        atualizacaoProduto.getEnvioPergunta().enviaEmail(produto, pergunta.getTitulo());
 
         //Definir tipo application/json
         final HttpHeaders httpHeaders= new HttpHeaders();
@@ -99,27 +92,5 @@ public class ProdutoAtualizacaoController {
 
         return ResponseEntity.ok().headers(httpHeaders).body(produto.getPerguntas().toString());
     }
-
-    private String montarEmailPergunta(Produto produto, PerguntaProduto pergunta){
-
-        String host = ServletUriComponentsBuilder.fromCurrentServletMapping().toUriString();
-        String linkProduto = String.format("%s/produtos/%d", host, produto.getId());
-
-        return String.format(
-                "Pergunta sobre o produto %d - %s\n\n" +
-                "Pergunta: %s\n\n" +
-                "Link do produto: %s",
-                produto.getId(), produto.getNome(),
-                pergunta.getTitulo(),
-                linkProduto);
-    }
-
-    private String montarAssuntoEmailPergunta(Produto produto) {
-        return String.format(
-                "Pergunta sobre o produto %d - %s",
-                produto.getId(), produto.getNome());
-    }
-
-
 
 }
