+++ src/main/java/br/com/zup/ecommerce/service/produto/AtualizacaoProduto.java
@@ -1,10 +1,8 @@
 package br.com.zup.ecommerce.service.produto;
 
 import br.com.zup.ecommerce.entities.produto.Produto;
-import br.com.zup.ecommerce.entities.produto.imagem.ImagensNovoRequest;
 import br.com.zup.ecommerce.entities.usuario.Usuario;
 import br.com.zup.ecommerce.security.UsuarioLogado;
-import br.com.zup.ecommerce.service.uploader.Uploader;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -12,7 +10,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import javax.persistence.EntityManager;
-import java.util.Set;
 
 /**
  * Contagem de carga intrínseca da classe: 7
@@ -23,7 +20,19 @@ public class AtualizacaoProduto {
 
     @Autowired
     //1
-    private Uploader uploader;
+    private AtualizacaoImagem atualizacaoImagem;
+
+    @Autowired
+    //1
+    private EnvioPergunta envioPergunta;
+
+    public AtualizacaoImagem getAtualizacaoImagem() {
+        return atualizacaoImagem;
+    }
+
+    public EnvioPergunta getEnvioPergunta() {
+        return envioPergunta;
+    }
 
     //1
     public Produto getProduto(Long id, EntityManager manager) {
@@ -48,10 +57,5 @@ public class AtualizacaoProduto {
         if(!produto.isDonoLogado(manager, usuario)) {
             throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         }
-    }
-
-    //1
-    public Set<String> enviaImagem(ImagensNovoRequest novasImagens) {
-        return uploader.enviaImagens(novasImagens.getImagens());
     }
 }
